Add SpinLock
## This Commit

Adds my idea of a `SpinLock`. I already read Chapter 4 a couple weeks
ago so the well is a bit poisoned. But I can't get the `loom` test to
pass so I've presumably forgotten something. In the next commit, I'll
implement what's in the book, and we'll see what I unlearned!