Update with learnings from book
## This Commit

Removes the explicit `Send` implementation on `SpinLock` because, as I
noted in my comment somewhat brainlessly, `UnsafeCell<T> impl Send where
T: Send` so this is automatically derived for `SpinLock`.

It also adds the test from the book which passes. The `loom` test still
doesn't so either it's a bug in `loom` or I've built the test
incorrectly.