Fix generic trait bounds in impl Sync for SpinLock
## This Commit

Updates the bounds on `T` for `SpinLock` to implement `Sync`. This
closes https://github.com/mlodato517/rust_atomics_and_locks/issues/1.

## Why?

When reviewing my code and comparing it with the book's implementation,
I noticed I had the incorrect bounds when implementing `Sync` for
`SpinLock`. I had required `T: Sync` when the book had `T: Send`.

I had erroneously thought that `Send` is only required when transferring
owned values to other threads. See the comments in the issue for more
details but I learned that `Send` is required to access `T`, `&T`, or
`&mut T` _at all_ on another thread. `Sync` is _only_ required if
concurrent access is required.