Update with learnings from Weak chapter
It looks like things were _mostly_ okay. Interesting to see `Relaxed`
ordering when increasing the count of `Arc`s when upgrading a `Weak`. I
guess that makes sense - there's already another `Arc` out there and
we're incrementing the value. It's not important that anything else has
"happened before" (except for the incrementing of the counter when that
other `Arc` was created but we don't need to worry about atomic
operations to the same value happening before - we're worried about
_other_ side effects).